Lots of tweaks to this
Using a different port "isOpen" check.

Set up cluster aware set to determine what ports are consumed by
other JVMs. There's also other methods to return use of the port.

I could probably encapsulate the port in an object that could check to
see if it's orphaned. Not sure how to do that, so I'm going to rely on
you to return a port to the 'pool.'